move and simplify some types
