+++ src/api/formatters.ts
@@ -11,7 +11,7 @@ import path from 'path'
 import { DEFAULT_CUCUMBER_PUBLISH_URL } from '../formatter/publish'
 import HttpStream from '../formatter/http_stream'
 import { Writable } from 'stream'
-import { IFormatterConfiguration } from '../configuration'
+import { IFormatterConfiguration } from './types'
 
 export async function initializeFormatters({
   cwd,
