+++ src/api/paths.ts
@@ -2,7 +2,7 @@ import { promisify } from 'util'
 import glob from 'glob'
 import path from 'path'
 import fs from 'mz/fs'
-import { ISourcesCoordinates } from '../configuration'
+import { ISourcesCoordinates } from './types'
 import { ISupportCodeCoordinates } from '../support_code_library_builder/types'
 
 export async function resolvePaths(
