+++ src/api/runCucumber.ts
@@ -8,8 +8,7 @@ import {
 } from '../cli/helpers'
 import { GherkinStreams } from '@cucumber/gherkin-streams'
 import PickleFilter from '../pickle_filter'
-import { IRunConfiguration } from '../configuration'
-import { IRunEnvironment, IRunResult } from './types'
+import { IRunConfiguration, IRunEnvironment, IRunResult } from './types'
 import { resolvePaths } from './paths'
 import { makeRuntime } from './runtime'
 import { initializeFormatters } from './formatters'
