+++ src/configuration/index.ts
@@ -1,5 +1,5 @@
 export { default as ArgvParser } from './argv_parser'
-export * from './convert_configuration'
+export * from '../api/convert_configuration'
 export * from './default_configuration'
 export * from './from_file'
 export * from './merge_configurations'
